refactor(cmd): move bridge service public-config builder out of run.go
Address PR review comment: run.go was already huge, so createBridgeService,
runningBridgeComponents and buildPublicConfig are extracted into their own
file, matching the naming already used by run_bridgeservice_test.go.
